関数化してコードをまとめた
